Implemented side selection and default imageview
