+++ corelib/src/main/java/dk/mustache/corelib/validation/failureview/FailureViewModel.kt
@@ -1,0 +1,12 @@
+package dk.mustache.corelib.validation.failureview
+
+import androidx.annotation.ColorInt
+import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
+import androidx.lifecycle.ViewModel
+
+class FailureViewModel(
+    @StringRes val message: Int,
+    @DrawableRes val imageResource: Int? = null,
+    @ColorInt val imageTint: Int?
+) : ViewModel()
