+++ corelib/src/main/res/layout/validation_view_failed_validation_textview.xml
@@ -4,8 +4,8 @@
     <data>
 
         <variable
-            name="validationMessage"
-            type="String" />
+            name="viewModel"
+            type="dk.mustache.corelib.validation.failureview.FailureViewModel" />
     </data>
 
     <androidx.appcompat.widget.AppCompatTextView xmlns:android="http://schemas.android.com/apk/res/android"
@@ -13,8 +13,7 @@
         android:id="@+id/txt_validation_failure"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="4dp"
         android:textColor="@color/red"
-        android:text="@{validationMessage}"
+        android:text="@{viewModel.message}"
         tools:text="Name must be filled out" />
 </layout>
