+++ corelib/src/main/res/values/styles.xml
@@ -97,8 +97,10 @@
     </style>
 
     <style name="Validation.EditText" parent="TextAppearance.AppCompat">
+        <item name="android:paddingStart">8dp</item>
         <item name="android:textColor">@drawable/validation_standard_textcolor</item>
         <item name="android:hint">@drawable/validation_standard_hintcolor</item>
+        <item name="android:imeOptions">actionNext</item>
     </style>
 
 </resources>
